Resolves #1. autoconfig.php has special handling during setup only
Investigating the Nextcloud source, can see SetupController only interested in a few variables from this first file to post that back through the installer. After that, the $CONFIG array more important in the other files.

Creating a special openshift.config.php file which brings in the desired variable into the $CONFIG array needed.